# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clinical Attending OS",
    "version": "0.2.0"
  },
  "paths": {
...
    "/submit_answer": {
      "post": {
        "summary": "Submit Answer",
        "operationId": "submitAnswer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/due_reviews": {
      "get": {
        "summary": "Due Reviews",
        "operationId": "getDueReviews",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Getduereviews"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/student_dashboard": {
      "get": {
        "summary": "Student Dashboard",
        "operationId": "getStudentDashboard",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDashboardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/ca1_coverage": {
      "get": {
        "summary": "Ca1 Coverage",
        "operationId": "getCA1Coverage",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CA1CoverageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/source_coverage": {
      "get": {
        "summary": "Source Coverage",
        "operationId": "getSourceCoverage",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceCoverageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/mark_mastered": {
      "post": {
        "summary": "Mastered",
        "operationId": "markMastered",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkTopicRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/mark_weak": {
      "post": {
        "summary": "Weak",
        "operationId": "markWeak",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkTopicRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/set_default_phase": {
      "post": {
        "summary": "Set Default Phase",
        "operationId": "setDefaultPhase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDefaultPhaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetDefaultPhaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/next_lesson": {
      "post": {
        "summary": "Next Lesson",
        "operationId": "nextLesson",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NextLessonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextLessonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/progress": {
      "get": {
        "summary": "Progress",
        "operationId": "getProgress",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/case_prep": {
      "post": {
        "summary": "Case Prep Endpoint",
        "operationId": "casePrep",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CasePrepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CasePrepResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/teaching_mode": {
      "post": {
        "summary": "Teaching Mode Endpoint",
        "description": "User teaches the GPT (prot\u00e9g\u00e9 effect). Returns probe seeds + grounded sources + rubric.",
        "operationId": "startTeachingMode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeachingModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeachingModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/follow_up": {
      "post": {
        "summary": "Follow Up Endpoint",
        "operationId": "followUp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/weak_patterns": {
      "get": {
        "summary": "Weak Patterns Endpoint",
        "operationId": "getWeakPatterns",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "window_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Window Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeakPatternsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "post": {
        "summary": "Search",
        "operationId": "searchSources",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AnswerRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": "voice"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "user_answer": {
            "type": "string",
            "title": "User Answer"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "phase": {
            "type": "string",
            "title": "Phase",
            "default": "warm_up_retrieval"
          },
          "subtopic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtopic"
          },
          "result": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "correct",
                  "partial",
                  "incorrect"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "mistake_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "recall",
                  "mechanism",
                  "prioritization",
                  "drug_dosing",
                  "monitoring",
                  "crisis_algorithm",
                  "oral_boards_framing",
                  "oral_presentation",
                  "incomplete_answer",
                  "overconfident_wrong",
                  "failure_to_escalate",
                  "differential_too_narrow",
                  "pretest_unstudied",
                  "critique_caught",
                  "critique_missed",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mistake Type"
          },
          "ideal_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ideal Answer"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "default": "medium"
          },
          "hints_used": {
            "type": "integer",
            "title": "Hints Used",
            "default": 0
          },
          "confidence_reported": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Reported"
          },
          "retrieval_sources": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Sources"
          },
          "source_citations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Citations"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "question",
          "user_answer",
          "topic"
        ],
        "title": "AnswerRequest"
      },
      "AnswerResponse": {
        "properties": {
          "evaluation": {
            "type": "string",
            "title": "Evaluation"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback"
          },
          "ideal_answer": {
            "type": "string",
            "title": "Ideal Answer"
          },
          "ideal_answer_key_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ideal Answer Key Points"
          },
          "mistake_type": {
            "type": "string",
            "title": "Mistake Type"
          },
          "updated_mastery": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated Mastery"
          },
          "next_review_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review Date"
          },
          "source_snippets": {
            "items": {
              "$ref": "#/components/schemas/SourceChunk"
            },
            "type": "array",
            "title": "Source Snippets"
          },
          "next_question": {
            "type": "string",
            "title": "Next Question"
          },
          "mini_teach": {
            "type": "string",
            "title": "Mini Teach",
            "default": ""
          },
          "teachback_prompt": {
            "type": "string",
            "title": "Teachback Prompt",
            "default": ""
          },
          "citation": {
            "type": "string",
            "title": "Citation",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "evaluation",
          "feedback",
          "ideal_answer",
          "mistake_type",
          "next_question"
        ],
        "title": "AnswerResponse"
      },
      "CA1CoverageResponse": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "default": "Stanford CA-1"
          },
          "granularity": {
            "type": "string",
            "title": "Granularity",
            "default": "fact"
          },
          "total_topics": {
            "type": "integer",
            "title": "Total Topics",
            "default": 0
          },
          "total_facts": {
            "type": "integer",
            "title": "Total Facts",
            "default": 0
          },
          "covered_topics": {
            "type": "integer",
            "title": "Covered Topics",
            "default": 0
          },
          "remaining_topics": {
            "type": "integer",
            "title": "Remaining Topics",
            "default": 0
          },
          "covered_facts": {
            "type": "integer",
            "title": "Covered Facts",
            "default": 0
          },
          "remaining_facts": {
            "type": "integer",
            "title": "Remaining Facts",
            "default": 0
          },
          "coverage_percent": {
            "type": "number",
            "title": "Coverage Percent",
            "default": 0.0
          },
          "next_targets": {
            "items": {
              "$ref": "#/components/schemas/CoverageTarget"
            },
            "type": "array",
            "title": "Next Targets"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/CoverageTarget"
            },
            "type": "array",
            "title": "Targets"
          }
        },
        "type": "object",
        "title": "CA1CoverageResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarkTopicRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "subtopic": {
            "type": "string",
            "title": "Subtopic",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "topic"
        ],
        "title": "MarkTopicRequest"
      },
...
      "SourceCoverageResponse": {
        "properties": {
          "granularity": {
            "type": "string",
            "title": "Granularity",
            "default": "fact"
          },
          "source_summaries": {
            "items": {
              "$ref": "#/components/schemas/SourceCoverageSummary"
            },
            "type": "array",
            "title": "Source Summaries"
          },
          "next_targets": {
            "items": {
              "$ref": "#/components/schemas/CoverageTarget"
            },
            "type": "array",
            "title": "Next Targets"
          }
        },
        "type": "object",
        "title": "SourceCoverageResponse"
      },
...
      "StudentDashboardResponse": {
        "properties": {
          "summary": {
            "items": {
              "$ref": "#/components/schemas/TopicSummary"
            },
            "type": "array",
            "title": "Summary"
          },
          "weakest_topics": {
            "items": {
              "$ref": "#/components/schemas/WeakTopic"
            },
            "type": "array",
            "title": "Weakest Topics"
          }
        },
        "type": "object",
        "title": "StudentDashboardResponse"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "servers": [
    {
      "url": "https://deanslaptop.tail712dd4.ts.net"
    }
  ]
}